pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gobin"
# version is handled automatically by setuptools_scm
dynamic = ["version"]
description = "AI Coding Agent CLI"
authors = [{ name = "Xgramm Platforms" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv",
    "rich",
    "prompt_toolkit",
    "google-genai",
]

[project.scripts]
gobin = "gobin.main:main"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "node-and-date"
